refactored Fit classes
