+++ autoarray/plot/mat_wrap/get_visuals.py
@@ -5,7 +5,7 @@ from autoarray.plot.mat_wrap.include import Include2D
 from autoarray.plot.mat_wrap.visuals import Visuals1D
 from autoarray.plot.mat_wrap.visuals import Visuals2D
 
-from autoarray.fit.fit_dataset import FitImaging
+from autoarray.fit.fit_imaging import FitImaging
 from autoarray.inversion.mappers.rectangular import MapperRectangularNoInterp
 from autoarray.inversion.mappers.voronoi import MapperVoronoiNoInterp
 from autoarray.mask.mask_2d import Mask2D
