+++ test_autoarray/inversion/mappers/test_abstract.py
@@ -133,3 +133,28 @@ def test__adaptive_pixel_signals_from___matches_util(grid_2d_7x7, image_7x7):
     )
 
     assert (pixel_signals == pixel_signals_util).all()
+
+
+def test__interpolated_array_from(grid_2d_7x7):
+
+    pixelization_grid_ndarray = aa.Grid2D.manual_slim(
+        [[0.1, 0.1], [1.1, 0.6], [2.1, 0.1], [0.4, 1.1], [1.1, 7.1], [2.1, 1.1]],
+        shape_native=(3, 2),
+        pixel_scales=1.0,
+    )
+
+    pixelization_grid = aa.Grid2DDelaunay(grid=pixelization_grid_ndarray)
+
+    mapper = aa.Mapper(
+        source_grid_slim=grid_2d_7x7, source_pixelization_grid=pixelization_grid
+    )
+
+    interpolated_array_via_mapper = mapper.interpolated_array_from(
+        values=np.array([1.0, 2.0, 3.0, 4.0, 5.0, 6.0]), shape_native=(3, 3)
+    )
+
+    interpolated_array_via_grid = pixelization_grid.interpolated_array_from(
+        values=np.array([1.0, 2.0, 3.0, 4.0, 5.0, 6.0]), shape_native=(3, 3)
+    )
+
+    assert (interpolated_array_via_mapper == interpolated_array_via_grid).all()
