+++ test_autoarray/test_preloads.py
@@ -4,7 +4,7 @@ import autoarray as aa
 
 from autoarray.mock.mock import MockMask
 from autoarray.mock.mock import MockDataset
-from autoarray.mock.mock import MockFit
+from autoarray.mock.mock import MockFitDataset
 from autoarray.mock.mock import MockMapper
 from autoarray.mock.mock import MockRegularization
 from autoarray.mock.mock import MockLEq
@@ -15,8 +15,8 @@ from autoarray.mock.mock import MockInversion
 #
 #     # fit inversion is None, so no need to bother with w_tilde.
 #
-#     fit_0 = MockFit(inversion=None)
-#     fit_1 = MockFit(inversion=None)
+#     fit_0 = MockFitDataset(inversion=None)
+#     fit_1 = MockFitDataset(inversion=None)
 #
 #     preloads = aa.Preloads(w_tilde=1, use_w_tilde=1)
 #     preloads.set_w_tilde_imaging(fit_0=fit_0, fit_1=fit_1)
@@ -26,10 +26,10 @@ from autoarray.mock.mock import MockInversion
 #
 #     # Noise maps of fit are different but there is an inversion, so we should not preload w_tilde and use w_tilde.
 #
-#     fit_0 = MockFit(
+#     fit_0 = MockFitDataset(
 #         inversion=1, noise_map=aa.Array2D.zeros(shape_native=(3, 1), pixel_scales=0.1)
 #     )
-#     fit_1 = MockFit(
+#     fit_1 = MockFitDataset(
 #         inversion=1, noise_map=aa.Array2D.ones(shape_native=(3, 1), pixel_scales=0.1)
 #     )
 #
@@ -49,8 +49,8 @@ from autoarray.mock.mock import MockInversion
 #
 #     dataset = MockDataset(psf=aa.Kernel2D.no_blur(pixel_scales=1.0), mask=mask)
 #
-#     fit_0 = MockFit(inversion=1, dataset=dataset, noise_map=noise_map)
-#     fit_1 = MockFit(inversion=1, dataset=dataset, noise_map=noise_map)
+#     fit_0 = MockFitDataset(inversion=1, dataset=dataset, noise_map=noise_map)
+#     fit_1 = MockFitDataset(inversion=1, dataset=dataset, noise_map=noise_map)
 #
 #     preloads = aa.Preloads(w_tilde=1, use_w_tilde=1)
 #     preloads.set_w_tilde_imaging(fit_0=fit_0, fit_1=fit_1)
@@ -72,8 +72,8 @@ def test__set_relocated_grid():
 
     # LEq is None so there is no mapper, thus preload mapper to None.
 
-    fit_0 = MockFit(inversion=None)
-    fit_1 = MockFit(inversion=None)
+    fit_0 = MockFitDataset(inversion=None)
+    fit_1 = MockFitDataset(inversion=None)
 
     preloads = aa.Preloads(relocated_grid=1)
     preloads.set_relocated_grid(fit_0=fit_0, fit_1=fit_1)
@@ -87,8 +87,8 @@ def test__set_relocated_grid():
         linear_obj_list=[MockMapper(source_grid_slim=2.0 * np.ones((3, 2)))]
     )
 
-    fit_0 = MockFit(inversion=MockInversion(leq=leq_0))
-    fit_1 = MockFit(inversion=MockInversion(leq=leq_1))
+    fit_0 = MockFitDataset(inversion=MockInversion(leq=leq_0))
+    fit_1 = MockFitDataset(inversion=MockInversion(leq=leq_1))
 
     preloads = aa.Preloads(relocated_grid=1)
     preloads.set_relocated_grid(fit_0=fit_0, fit_1=fit_1)
@@ -100,8 +100,8 @@ def test__set_relocated_grid():
     leq_0 = MockLEq(linear_obj_list=[MockMapper(source_grid_slim=np.ones((3, 2)))])
     leq_1 = MockLEq(linear_obj_list=[MockMapper(source_grid_slim=np.ones((3, 2)))])
 
-    fit_0 = MockFit(inversion=MockInversion(leq=leq_0))
-    fit_1 = MockFit(inversion=MockInversion(leq=leq_1))
+    fit_0 = MockFitDataset(inversion=MockInversion(leq=leq_0))
+    fit_1 = MockFitDataset(inversion=MockInversion(leq=leq_1))
 
     preloads = aa.Preloads(relocated_grid=1)
     preloads.set_relocated_grid(fit_0=fit_0, fit_1=fit_1)
@@ -113,8 +113,8 @@ def test__set_linear_obj_list():
 
     # LEq is None so there is no mapper, thus preload mapper to None.
 
-    fit_0 = MockFit(inversion=None)
-    fit_1 = MockFit(inversion=None)
+    fit_0 = MockFitDataset(inversion=None)
+    fit_1 = MockFitDataset(inversion=None)
 
     preloads = aa.Preloads(linear_obj_list=1)
     preloads.set_linear_obj_list(fit_0=fit_0, fit_1=fit_1)
@@ -126,8 +126,8 @@ def test__set_linear_obj_list():
     leq_0 = MockLEq(linear_obj_list=[MockMapper(mapping_matrix=np.ones((3, 2)))])
     leq_1 = MockLEq(linear_obj_list=[MockMapper(mapping_matrix=2.0 * np.ones((3, 2)))])
 
-    fit_0 = MockFit(inversion=MockInversion(leq=leq_0))
-    fit_1 = MockFit(inversion=MockInversion(leq=leq_1))
+    fit_0 = MockFitDataset(inversion=MockInversion(leq=leq_0))
+    fit_1 = MockFitDataset(inversion=MockInversion(leq=leq_1))
 
     preloads = aa.Preloads(linear_obj_list=1)
     preloads.set_linear_obj_list(fit_0=fit_0, fit_1=fit_1)
@@ -139,8 +139,8 @@ def test__set_linear_obj_list():
     leq_0 = MockLEq(linear_obj_list=[MockMapper(mapping_matrix=np.ones((3, 2)))])
     leq_1 = MockLEq(linear_obj_list=[MockMapper(mapping_matrix=np.ones((3, 2)))])
 
-    fit_0 = MockFit(inversion=MockInversion(leq=leq_0))
-    fit_1 = MockFit(inversion=MockInversion(leq=leq_1))
+    fit_0 = MockFitDataset(inversion=MockInversion(leq=leq_0))
+    fit_1 = MockFitDataset(inversion=MockInversion(leq=leq_1))
 
     preloads = aa.Preloads(linear_obj_list=1)
     preloads.set_linear_obj_list(fit_0=fit_0, fit_1=fit_1)
@@ -162,8 +162,8 @@ def test__set_linear_obj_list():
         ]
     )
 
-    fit_0 = MockFit(inversion=MockInversion(leq=leq_0))
-    fit_1 = MockFit(inversion=MockInversion(leq=leq_1))
+    fit_0 = MockFitDataset(inversion=MockInversion(leq=leq_0))
+    fit_1 = MockFitDataset(inversion=MockInversion(leq=leq_1))
 
     preloads = aa.Preloads(linear_obj_list=1)
     preloads.set_linear_obj_list(fit_0=fit_0, fit_1=fit_1)
@@ -179,8 +179,8 @@ def test__set_operated_mapping_matrix_with_preloads():
 
     # LEq is None thus preload it to None.
 
-    fit_0 = MockFit(inversion=None)
-    fit_1 = MockFit(inversion=None)
+    fit_0 = MockFitDataset(inversion=None)
+    fit_1 = MockFitDataset(inversion=None)
 
     preloads = aa.Preloads(
         operated_mapping_matrix=1,
@@ -206,8 +206,8 @@ def test__set_operated_mapping_matrix_with_preloads():
     leq_0 = MockLEqImaging(blurred_mapping_matrix=blurred_mapping_matrix_0)
     leq_1 = MockLEqImaging(blurred_mapping_matrix=blurred_mapping_matrix_1)
 
-    fit_0 = MockFit(inversion=MockInversion(leq=leq_0))
-    fit_1 = MockFit(inversion=MockInversion(leq=leq_1))
+    fit_0 = MockFitDataset(inversion=MockInversion(leq=leq_0))
+    fit_1 = MockFitDataset(inversion=MockInversion(leq=leq_1))
 
     preloads = aa.Preloads(
         operated_mapping_matrix=1,
@@ -231,8 +231,8 @@ def test__set_operated_mapping_matrix_with_preloads():
         curvature_matrix_counts=curvature_matrix_counts,
     )
 
-    fit_0 = MockFit(inversion=inversion_0)
-    fit_1 = MockFit(inversion=MockInversion(leq=leq_1))
+    fit_0 = MockFitDataset(inversion=inversion_0)
+    fit_1 = MockFitDataset(inversion=MockInversion(leq=leq_1))
 
     preloads = aa.Preloads(
         operated_mapping_matrix=1,
@@ -256,8 +256,8 @@ def test__set_regularization_matrix_and_term():
 
     # Inversion is None thus preload log_det_regularization_matrix_term to None.
 
-    fit_0 = MockFit(inversion=None)
-    fit_1 = MockFit(inversion=None)
+    fit_0 = MockFitDataset(inversion=None)
+    fit_1 = MockFitDataset(inversion=None)
 
     preloads = aa.Preloads(log_det_regularization_matrix_term=1)
     preloads.set_regularization_matrix_and_term(fit_0=fit_0, fit_1=fit_1)
@@ -267,12 +267,12 @@ def test__set_regularization_matrix_and_term():
 
     # Inversion's log_det_regularization_matrix_term are different thus no preloading.
 
-    fit_0 = MockFit(
+    fit_0 = MockFitDataset(
         inversion=MockInversion(
             log_det_regularization_matrix_term=0, regularization_list=[regularization]
         )
     )
-    fit_1 = MockFit(
+    fit_1 = MockFitDataset(
         inversion=MockInversion(
             log_det_regularization_matrix_term=1, regularization_list=[regularization]
         )
@@ -288,7 +288,7 @@ def test__set_regularization_matrix_and_term():
 
     preloads = aa.Preloads(log_det_regularization_matrix_term=2)
 
-    fit_0 = MockFit(
+    fit_0 = MockFitDataset(
         inversion=MockInversion(
             leq=MockLEq(linear_obj_list=[MockMapper()]),
             log_det_regularization_matrix_term=1,
@@ -296,7 +296,7 @@ def test__set_regularization_matrix_and_term():
             preloads=preloads,
         )
     )
-    fit_1 = MockFit(
+    fit_1 = MockFitDataset(
         inversion=MockInversion(
             leq=MockLEq(linear_obj_list=[MockMapper()]),
             log_det_regularization_matrix_term=1,
